backend/backendapi/requirements.txt
asgiref==3.3.4
Django==3.2.4
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
Markdown==3.3.4
mysql-connector-python==8.0.25
pytz==2021.1
sqlparse==0.4.1
